refactor(plugin): keep compatibility changes within lint budgets
Move Claude root-instruction helpers out of the oversized parser file and keep marketplace fixtures line-neutral while preserving the official skill-based coverage.

Verification: go run ./tools/repolint; go test ./...; ./scripts/cache-guard.sh.
