+++ internal/pluginpkg/pluginpkg.go
@@ -655,27 +655,6 @@ func ManifestPaths() []string {
 	return []string{NativeManifest, CodexManifest, ClaudeManifest}
 }
 
-func applyClaudeCompatibility(root string, manifest *Manifest) ([]string, []CompatibilityIssue) {
-	appendRootClaudeInstructions(root, manifest)
-	return appendClaudeCompatibility(root, manifest)
-}
-
-func appendRootClaudeInstructions(root string, manifest *Manifest) {
-	path := filepath.Join(root, claudeInstructions)
-	info, err := os.Stat(path)
-	if err != nil || !info.Mode().IsRegular() {
-		return
-	}
-	if manifest.Hooks == nil {
-		manifest.Hooks = map[string][]Hook{}
-	}
-	manifest.Hooks["SessionStart"] = append(manifest.Hooks["SessionStart"], Hook{
-		ContextFile: claudeInstructions,
-		Cwd:         ".",
-		Description: "Plugin CLAUDE.md startup context from " + manifest.Name,
-	})
-}
-
 func claudeTimeoutMillis(seconds int) int {
 	if seconds <= 0 {
 		return 0
